Compare by ID + Similarity Coefficient
Completed the compare by ID functionality so it through the YAML one may se what the key columns for each side of the Bidirectional comparison.

Full priority is given for full matches over the key columns.

TODO: Fix penalization for not including information under certain columns. It has been analyzed that reference entries that have more columns filled are most likely to be matched in average with any random entry.